+++ packages/nano-ui/lib/NanoUI/Input.hs
@@ -27,6 +27,7 @@ module NanoUI.Input
 
     -- * Cursors
   , UiCursorKind (..)
+  , CursorShape
   , grabHoverKind
   , grabDragKind
   ) where